cast the enum the trace call formats as an integer
DbgTrc is a fmt::make_printf_args wrapper now, and fmt is type safe where a
vararg was not: an unscoped enum has no formatter, so a %d that used to promote
it to int is a compile error naming type_is_unformattable_for. static_cast<int>
is how the rest of the tree writes this, AILogic/PlayerReinforcement.cpp and
Scripts.cpp among others.

With this MapEditorLib compiles and links.

+++ Versions/Temporary/Engine/Sources/MapEditorLib/FolderController.cpp
@@ -155,7 +155,7 @@ bool CFolderController::RedoWithoutUpdateViews()
 		}
 		dwTime = GetCurrentTimeMilliseconds() - dwTime;
 		DebugTrace( "CFolderController: operation:%d, type: <%s>, dest:<%s>, source:<%s>, result:%s, time: %dmc",
-								itUndoData->eType,
+								static_cast<int>( itUndoData->eType ),
 								GetObjectSet().szObjectTypeName.c_str(),
 								itUndoData->szDestination.c_str(),
 								itUndoData->szSource.c_str(),
